Simplify the core and define a set of "Boucle operation" primitives
Rather than trying to control the core with MIDI messages, we should
use a custom protocol that maps to the actual loop operations that the
core is doing. We still need a MIDI mapping at some level, but this
approach allows the MIDI mapping to change without affecting the
core code.

This also means the core doesn't need to be synced to the tempo. Layers
on top (such as mapping to MIDI notes) may want to allow triggering
beat-synced events, and will thus need to know the tempo, but the core
now receives all events in frames.

This may change again when I try to add support for syncing to MIDI
timecode and MIDI machine code (hopefully that can happen in the
plugin host layer instead, but JACK's transport explicitly doesn't
support MIDI timecode to avoid forcing clients to vary the samplerate.)