core: Ops now transform absolute play clock
The previous approach worked when operations were used in isolation,
but they couldn't be combined correctly because information was lost
between each stage. Now each operation returns a transform that will
be applied to the play clock. The engine itself is responsible
for calculating the final value, and only once all transforms are
applied we wrap it to be within the actual loop buffer.

This is hopefully all we need to start implementing a PianoController
that can trigger ops based on MIDI inputs.