Use Instant and Duration everywhere
To ensure accurate timings, we now use the system clock right up until the last minute.
For each sample we convert the clock into samples.
This isn't the most efficient approach but should be the most accurate, we can optimize later.

For batch mode, we'll need to make sure the batch start time matches the start time of the engine,
which can be done if we allow overriding time when processing the buffer
I guess.