Initial live input manipulation!!
I abandoned the 'everything synced to wall clock' approach because
of glitches in the sound - it seems I can't get this approach
sample-perfect. Use samples everywhere, which will only go wrong
if we have buffer underruns.

Various effects are a bit glitchy, maybe bugs or just the expected
lack of interpolation as we jump around.