organelle: Use JACK
CPAL cannot be used on Organelle.

  * Using default audio device gives a EPIPE error in snd_pcm_start()
  * Specifying device `default=CARD=Codec` fails because of
    https://github.com/RustAudio/cpal/issues/615
  * Using JACK backend fails for reasons I didn't investigate - the
    `feedback` example fails with 'unsupported configuration' even
    though it's clearly using a supported configuration.

JACK is available on Organelle and can play sound, we just need to start
it in the run.sh script.

I don't have time to fix any of the CPAL problems so let's switch to
JACK for everything.

This can also be tested locally on desktop, but see:
https://github.com/RustAudio/rust-jack/issues/142